Due to pm2 start errors move server to root and add dotenv in various files
