+++ daily_booking_forecast/step_1_get_booking_by_hour_data.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 const path = require('path');
 const dotenv = require('dotenv');
-dotenv.config();
+dotenv.config({ path: "../.env" });
 
 const mysql = require('mysql2');
 const { Client } = require('ssh2');
